Update the attributes of bscui elements in 'shiny' app

Update the attributes of bscui elements in 'shiny' app

Usage

update_bscui_attributes(
  proxy,
  element_attributes,
  to_ignore = NULL,
  targeted_tags = NULL
)

Arguments

proxy

a bscui_Proxy object

element_attributes

a data frame with an id column providing the element identifier and one column per attribute name.

to_ignore

of elements to ignore: if those elements are children of elements to update they won't be updated. This parameter is not taken into account when there is no "id" column in the element_styles data frame.

targeted_tags

affected tag names. If NULL (default), the structure_shapes of the bscui object

Value

the provided proxy object

Examples

if(interactive()){
   from_shiny <- new.env()
   shiny::runApp(system.file(
      "examples", "shiny-anatomogram", package = "bscui"
   ))
   for(n in names(from_shiny)){
      bscui(from_shiny[[n]]) |> print()
   }
}